ASO_Chat_Backend/app/requirements.txt
Django==3.2.9
gunicorn==20.1.0
psycopg2-binary==2.9.1
djangorestframework
django-cors-headers
